% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.kron.names.r
\name{clean.kron.names}
\alias{clean.kron.names}
\title{clean.kron.names}
\usage{
clean.kron.names(xnames)
}
\arguments{
\item{xnames}{a character vector of rownames or colnames}
}
\value{
a character vector of cleaned up names
}
\description{
Clean up rownames and colnames returned by \code{kronecker}
}
\examples{
    A <- matrix(1:4, ncol=2)
    B <- matrix(5:8, ncol=2)
    rownames(A) <- colnames(A) <- c("a1", "a2")
    rownames(B) <- colnames(B) <- c("b1", "b2")
    C <- kronecker(A,B, make.dimnames = TRUE)
    clean.kron.name(colnames(C))
    
}
