% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.terms.r
\name{collect.terms}
\alias{collect.terms}
\title{collect.terms}
\usage{
collect.terms(A)
}
\arguments{
\item{A}{a matrix, with rownames and colnames}
}
\value{
the matrix with redundant columns dropped and redundant
    rows summed
}
\description{
Eliminate redundant terms in the column space 
    and collect redundant terms in the row space
}
\examples{

}
