% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colSds.r
\name{colSds}
\alias{colSds}
\title{colSds}
\usage{
colSds(x, na.rm = FALSE)
}
\arguments{
\item{x}{a matrix or data frame}

\item{na.rm}{logical Should missing values be removed?}
}
\value{
a named vector of standard devations
}
\description{
Column standard deviations for matrices and data frames
}
\examples{
    colSds(mtcars[,c("wt", "disp")])
    
}
\seealso{
\code{colMeans}
}
