% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.all.terms.r
\name{generate.all.terms}
\alias{generate.all.terms}
\title{generate.all.terms}
\usage{
generate.all.terms(tvd)
}
\arguments{
\item{tvd}{a terms/variables/degrees table, see \code{terms.vars.degrees}}
}
\value{
a character vector of model terms
}
\description{
Generate a complete list of terms where some lower order terms
    may have been dropped
}
\examples{
    fit <- lm(mpg ~ 0 + wt*disp, data=mtcars)
    generate.all.terms(terms.vars.degrees(names(coef(fit))))

}
