% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kron.r
\name{kron}
\alias{kron}
\title{kron}
\usage{
kron(A, B)
}
\arguments{
\item{A, B}{matrices that generate a direct product}
}
\value{
the matrix \code{kronecker(A,B, make.dimnames=TRUE)}, but
    with the names formed in reverse order, and \code{(Intercept)}
    terms cleaned up.
}
\description{
Form Kronecker (direct) products with useful row 
    and column names
}
\examples{
    A <- matrix(1:4, ncol=2)
    B <- matrix(5:8, ncol=2)
    rownames(A) <- colnames(A) <- c("a1", "a2")
    rownames(B) <- colnames(B) <- c("b1", "b2")
    kronecker(A,B, make.dimnames = TRUE)
    kron(A,B)
    
}
\seealso{
\code{clean.kron.names}, \code{kronecker}
}
