% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.to.matrix.r
\name{mean.to.matrix}
\alias{mean.to.matrix}
\title{mean.to.matrix}
\usage{
\method{mean}{to.matrix}(x)
}
\arguments{
\item{x}{a named scalar, the mean}
}
\value{
a matrix
}
\description{
Convert a scalar mean to a coefficient recentering matrix
}
\examples{
    mu <- 3
    names(mu) <- "x"
    mean.to.matrix(mu)
    
}
\seealso{
\code{sd.to.matrix}
}
