% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastex.r
\name{pastex}
\alias{pastex}
\title{pastex}
\usage{
pastex(mat)
}
\arguments{
\item{mat}{a matrix}
}
\value{
a character vector
}
\description{
Paste together the rows of a matrix, separated by a colon
}
\examples{
    A <- matrix(letters[1:6], ncol=3)
    pastex(A)
    
}
