% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly.to.interaction.r
\name{poly.to.interaction}
\alias{poly.to.interaction}
\title{poly.to.interaction}
\usage{
poly.to.interaction(terms)
}
\arguments{
\item{terms}{a character vector of simple terms
(no intercation terms)}
}
\value{
a character vector of converted terms
}
\description{
Converts polynomial terms like I(x^2)
    to the interaction form x:x
}
\examples{
    poly.to.interaction(c("Intercept", "x", "I(x^2)"))
    
}
\seealso{
\code{poly.fix}
}
