% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyterm.r
\name{polyterm}
\alias{polyterm}
\title{polyterm}
\usage{
polyterm(mu, b.terms, type = "center")
}
\arguments{
\item{mu}{a vector of recentering or rescaling constants}

\item{b.terms}{a character vector of terms}

\item{type}{either \code{center} or \code{scale}}
}
\value{
a recentering or rescaling matrix
}
\description{
Generate recentering and rescaling matrices
    for polynomial terms
}
