% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recentering.matrix.r
\name{recentering.matrix}
\alias{recentering.matrix}
\title{recentering.matrix}
\usage{
recentering.matrix(x, terms, type = "center")
}
\arguments{
\item{x}{named vector of recentering or rescaling constants, one
per variable in \code{terms}}

\item{terms}{character vector of terms (coefficients) in a model}

\item{type}{the type of matrix to return, either \code{center} 
(the default) or \code{scale}}
}
\value{
a recentering or rescaling matrix
}
\description{
Recentering and rescaling matrix generation
}
\examples{
    fit <- lm(mpg ~ wt*disp, data=mtcars)
    mu    <- colMeans(mtcars[,c("wt", "disp")])
    sigma <- colSds(mtcars[,c("wt", "disp")])
    terms <- names(coef(fit))
    
    recentering.matrix(mu, terms)
    recentering.matrix(mu, terms, type="scale")
    
}
