% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.to.gm.r
\name{ref.to.gm}
\alias{ref.to.gm}
\title{ref.to.gm}
\usage{
ref.to.gm(k)
}
\arguments{
\item{k}{the number of levels in the factor
to be converted}
}
\value{
a coeficient recentering matrix (the inverse of
    the data recentering matrix)
}
\description{
Reference-to-grand-mean conversion matrix
}
\examples{
    ref.to.gm(3)        # coefficient recentering
    solve(ref.to.gm(3)) # data recentering/recoding
 
}
\seealso{
\code{mean.to.matrix}, \code{sd.to.matrix}
}
