% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.to.matrix.r
\name{sd.to.matrix}
\alias{sd.to.matrix}
\title{sd.to.matrix}
\usage{
sd.to.matrix(x)
}
\arguments{
\item{x}{a named scalar, the standard devation}
}
\value{
a matrix
}
\description{
Convert a scalar standard devation
    to a coefficient rescaling matrix
}
\examples{
    sigma <- 3
    names(sigma) <- "x"
    sd.to.matrix(sigma)
    
}
\seealso{
\code{mean.to.matrix}
}
